feat(Editor): add WCo2Editor component for enhanced text editing capabilities feat(Menu): introduce Wco2Menu component for dynamic menu rendering chore(index): update exports to include new Editor and Menu components refactor(CatalogueCard): remove console log for cleaner code refactor(menu): remove unnecessary console log for improved performance The WCo2Editor and Wco2Menu components are added to enhance the text editing and menu functionalities, respectively. The index files are updated to export these new components. Console log statements are removed from CatalogueCard and menu files to maintain cleaner code and improve performance.
