+++ platform/web/packages/components/src/Editor/WCo2Editor.tsx
