+++ platform/web/packages/components/src/index.ts
@@ -22,8 +22,8 @@ export * from "./SubtitledLinkTicket"
 export * from "./SearchFilter"
 export * from "./Tabs"
 export * from "./LexicalEditor"
-export * from "./WCo2Editor"
+export * from "./Editor"
 export * from "./i18n"
 export * from "./LanguageSelector"
-export * from "./Wco2Menu"
+export * from "./Menu"
