+++ platform/web/packages/web-app/src/App/menu/index.tsx
@@ -140,7 +140,6 @@ export const maybeAddItem = <T,>(condition: boolean, item: T): T[] => {
 const mapCatalogueRef = (currentPaths: string[], cataloguesAll: (tab?: string, ...objectIds: string[]) => string) => (item: CatalogueRefTree): MenuItem => {
   const newPath = [...currentPaths, item.identifier]
   const ref = cataloguesAll(item.identifier, ...currentPaths)
-  console.log("ref", ref)
   return {
     key: item.identifier,
     to: ref,
